+++ src/msw/dcprint.cpp
@@ -361,7 +361,7 @@ WXHDC WXDLLEXPORT wxGetPrinterDC(const wxPrintData& printDataConst)
 
     LPDEVMODE lpDevMode = (LPDEVMODE) NULL;
 
-    HGLOBAL hDevMode = (HGLOBAL) printData.GetNativeData();
+    HGLOBAL hDevMode = (HGLOBAL)(DWORD) printData.GetNativeData();
 
     if ( hDevMode )
         lpDevMode = (DEVMODE*) GlobalLock(hDevMode);
